reorganizing market pi tests
